# Conflicts:
#	CatchMate/presentation/src/main/res/navigation/nav_graph.xml

+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/user/GetBlockedUserListResponse.kt
@@ -3,7 +3,7 @@ package com.catchmate.domain.model.user
 data class GetBlockedUserListResponse(
     val userInfoList: List<GetUserProfileResponse>,
     val totalPages: Int,
-    val totalElements: Long,
+    val totalElements: Int,
     val isFirst: Boolean,
     val isLast: Boolean,
 )
